pyproject.toml
[tool.poetry]
name = "classified"
version = "0.1.0"
description = ""
authors = ["Vik Paruchuri <vik.paruchuri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
python-dotenv = "^1.0.0"
openai = "^0.28.1"
tiktoken = "^0.5.1"
tenacity = "^8.2.3"
sqlmodel = "^0.0.8"
numpy = "^1.26.1"
alembic = "^1.12.0"
datasets = "^2.14.5"
tqdm = "^4.66.1"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
